pyproject.toml
[project]
name = "Pulsepost-v2"
version = "0.1.0"
description = "A LangChain agent that discovers trending topics from web/X/Reddit, presents ranked topic candidates to the user, fetches and extracts content for the selected topic, generates a LinkedIn post from that content via an LLM, and posts to LinkedIn via the LinkedIn API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "ddgs>=9.5.2",
    "google-search-results>=2.4.2",
    "langchain>=1.0.0a13",
    "langchain-google-genai>=2.1.12",
    "langgraph>=1.0.0a4",
    "praw>=7.8.1",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "streamlit>=1.50.0",
    "trafilatura>=2.0.0",
]
